Change to work around an Adobe software issue (as outlined in Issue #15).
We now cache the contentType in Document in order to avoid calling the Adobe version (because we can't be sure whether it will work or whether it will throws an exception).  We also no longer set the length in the response coming from rest-services.server.GeneratePdfOutput because we can't reliably call Document.getContentLength() either.

I've added tests into it.tests to test with a PDF that triggers this stupid behaviour in the Adobe APIs.