Removed the ability to pass in URLs or CRX Urls to the service since it doesn't make sense to do so.  All templates must be relative to a particular location in CRX.
I could have perhaps removed the ability to pass in a PathOrUrl completely in favour of just allowing Path or String, however I decided to remain consistent with other APIs.  This may be a mistake and may cause some confusion for users of the library, but there you have it.  We can always deprecate it later I suppose.