Replaced String constructor with URL constructor.
This means that we don't have to generate a MalformedURL exception.
This pushes the requirement to handle a poorly formed URL up into the
upper levels of the code which is better.  It will be up to the caller
to make an URL object (which they can do closer to the point where the
string is acquired).