Improved implementation of PrintConfigImpl
Encountered an integration test issue with the way PrintConfigImpl was
implemented (it was implemented as a class with many constants).
Changed it to an enum to fix the issue.

The issue was that the rest-services client code serialized the
PrintConfig with toString().  Since PrintConfigImpl was implemented as a
class, you get an object handle (which is not good for deserializing on
the rest-services server code.  By re-implementing it as an enum ,
toString returns something meaningful that can be recreated on the
server side.

The downside of this approach is that custom printer configs are not
supported at this time.  They could perhaps be supported later with a
bunch more code (to serialize everything into a String that can be
re-constituted on the server).  For the time being we just produce a
warning that it is not supported.