Updated fluentforms library version to 0.0.3-SNAPSHOT.
This is in preparation for (maybe) creating a 0.0.2 release that
utilizes jersey 2.x instead of jersey 3.x.

We have a client that may need to stay on Spring Boot 2.x which implies
using Jersey 2.x.  We were perhaps a bit hasty in upgrading this library
to Jersey 3.x without updating the version number.  I am looking at
retroactively doing that.