Added a "client prefix" to the filter builder.
The client prefix gets applied in front of the "/aem" prefix.  This is
useful if the client wants (or needs) to nest the aem urls within
another location (e.g. if the client code is in a .war that will
deployed to an app server, then that war may have all the endpoints in
it prefixed with the war name).