✨ Implemented TestContainers startup
This is the initial first step in getting a more self-contained
integration test.  The code now starts up a docker container containing
AEM (assuming that one is available).

Next step is to deploy sample files to the container before runnning the
tests.

The code assumes that an AEM container is available (in this case one
that is hosted by the 4PointSolutions-PS GitHub org.

The container code can be turned off by a constant in
AbstractAemContainerTest.  When turned off, the behaviour reverts to
previous behaviour (i.e. it assumes a copy of AEM is running locally),