✅✨ Added support for using TestContainers for running IT tests (WIP)
Integration tests can now be configured to use AEM running as a Docker
container (via TestContainers).

This is still work in progress, but it's breathing.  The container image
I'm using doesn't have the samples files deployed in it.  It would need
this in order for the tests to work.  See the pom.xml for details on
what needs to be configured for the IT tests to work.