♻️ Modified services to explicitly use AdobeDocumentFactoryImpl
The rest-services code used DocumentFactory.getDefault() which obscured
which implementation was being used (even though it was always
AdobeDocumentFactoryImpl).  Since we know that rest-services will be
running on an AEM server (and will therefore always have
AdobeDocumentImpl available) we might as well call it explicitly so that
it's obvious which implementation is being used.