✨ Completed instructions for integration testing container
Created and tested instructions for creating an integration testing
container that allows all the integration tests that run by default to
pass.  The container image has all the necessary sample files placed on
disk and all the necessary sample files deployed to AEM.

In theory, this would allow the integration tests to be run as part of a
GitHub action if the test container image were available however because
AEM is proprietary, an AEM image cannot be placed in a publicly
accessible location.