✅ Updated WireMock mappings to match the response from AEM.
Originally, the code used an Adaptive Form recording, but when the
TestContainers code was created, it used an HTML5 form instead.  At that
time, the WireMock recordings weren't updated so the WireMock tests
still used the Adaptive Form.

Now both WireMock and TestContainers tests use the same HTML5 form and
so the same container image (running on port 4502) can be used to
re-create the WireMock recordings.