✨ Got things compiling in IDE with null checking on
Still lots of warnings but no errors.  Not sure if it is worth the
effort to get rid of the warnings.  Most of the warnings are about JDK
classes (Eclipse doesn't seem to have any built in mechanism for this -
it relies on outside libraries which are out of date). I think the
Eclipse null checking may need some work.