➖ Removed dependency on Spring RestClient
The testing classpath used to include the Spring classes which triggered
the Spring autoconfiguration code (which, in turn, required that we mock
out some of the classes provided by Spring that were used by the spring
mvc autoconfiguration).

Now that the testing classes no longer include those dependencies, we
don't need to mock them.